宾馆项目-小吧入账和商品增删改查完毕

+++ src/main/java/com/ninja/wangjia_backend/model/entity/ShopOrder.java
@@ -28,7 +28,7 @@ public class ShopOrder implements Serializable {
     /**
      * 商品单价
      */
-    private BigDecimal shopPrice;
+    private Double shopPrice;
 
     /**
      * 顾客类型(0:在住,1:外来)
